requirements.txt
absl-py==2.1.0
Adafruit-Blinka==8.24.0
adafruit-circuitpython-ads1x15==2.2.24
adafruit-circuitpython-bh1750==1.1.9
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.3
adafruit-circuitpython-register==1.9.17
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.53.0
Adafruit-PureIO==1.1.11
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==20181003210633
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
jax==0.4.28
keras==2.12.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
mock==5.1.0
namex==0.0.8
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.1.3
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyftdi==0.55.0
Pygments==2.18.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
pyusb==1.2.1
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
schedule==1.2.1
scipy==1.13.1
six==1.16.0
sysv-ipc==1.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow @ file:///home/kaboom/Bot/tensorflow-2.12.0rc0-cp311-none-linux_aarch64.whl#sha256=03eb85057e96c94bb78385b8472acb1918e779c8a6b50682350facc5d2887d9f
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tflite-runtime==2.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.14.1
